Fix issue psalm.
