+++ src/Query/QueryBuilder.php
@@ -76,7 +76,7 @@ use function trim;
  * @property string[] $expressionBuilders Array of builders that should be merged with the pre-defined one's in
  * {@see expressionBuilders} property. This property is write-only.
  */
-class QueryBuilder
+abstract class QueryBuilder
 {
     /**
      * The prefix for automatically generated query binding parameters.
@@ -134,6 +134,104 @@ class QueryBuilder
         $this->expressionBuilders = $this->defaultExpressionBuilders();
         $this->conditionClasses = $this->defaultConditionClasses();
     }
+
+    /**
+     * Creates a SQL command for adding a default value constraint to an existing table.
+     *
+     * @param string $name the name of the default value constraint.
+     * The name will be properly quoted by the method.
+     * @param string $table the table that the default value constraint will be added to.
+     * The name will be properly quoted by the method.
+     * @param string $column the name of the column to that the constraint will be added on.
+     * The name will be properly quoted by the method.
+     * @param mixed $value default value.
+     *
+     * @throws Exception|NotSupportedException if this is not supported by the underlying DBMS.
+     *
+     * @return string the SQL statement for adding a default value constraint to an existing table.
+     */
+    abstract public function addDefaultValue(string $name, string $table, string $column, $value): string;
+
+    /**
+     * Builds a SQL statement for enabling or disabling integrity check.
+     *
+     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema.
+     * @param string $table the table name. Defaults to empty string, meaning that no table will be changed.
+     * @param bool $check whether to turn on or off the integrity check.
+     *
+     * @throws Exception|NotSupportedException if this is not supported by the underlying DBMS.
+     *
+     * @return string the SQL statement for checking integrity.
+     */
+    abstract public function checkIntegrity(string $schema = '', string $table = '', bool $check = true): string;
+
+    /**
+     * Creates a SQL command for dropping a default value constraint.
+     *
+     * @param string $name the name of the default value constraint to be dropped.
+     * The name will be properly quoted by the method.
+     * @param string $table the table whose default value constraint is to be dropped.
+     * The name will be properly quoted by the method.
+     *
+     * @throws Exception|NotSupportedException if this is not supported by the underlying DBMS.
+     *
+     * @return string the SQL statement for dropping a default value constraint.
+     */
+    abstract public function dropDefaultValue(string $name, string $table): string;
+
+    /**
+     * Creates a SQL statement for resetting the sequence value of a table's primary key.
+     *
+     * The sequence will be reset such that the primary key of the next new row inserted will have the specified value
+     * or 1.
+     *
+     * @param string $tableName the name of the table whose primary key sequence will be reset.
+     * @param array|int|string|null $value the value for the primary key of the next new row inserted. If this is not
+     * set, the next new row's primary key will have a value 1.
+     *
+     * @throws Exception|NotSupportedException if this is not supported by the underlying DBMS.
+     *
+     * @return string the SQL statement for resetting sequence.
+     */
+    abstract public function resetSequence(string $tableName, array|int|string|null $value = null): string;
+
+    /**
+     * Creates an SQL statement to insert rows into a database table if they do not already exist (matching unique
+     * constraints), or update them if they do.
+     *
+     * For example,
+     *
+     * ```php
+     * $sql = $queryBuilder->upsert('pages', [
+     *     'name' => 'Front page',
+     *     'url' => 'http://example.com/', // url is unique
+     *     'visits' => 0,
+     * ], [
+     *     'visits' => new \Yiisoft\Db\Expression('visits + 1'),
+     * ], $params);
+     * ```
+     *
+     * The method will properly escape the table and column names.
+     *
+     * @param string $table the table that new rows will be inserted into/updated in.
+     * @param array|Query $insertColumns the column data (name => value) to be inserted into the table or instance
+     * of {@see Query} to perform `INSERT INTO ... SELECT` SQL statement.
+     * @param array|bool $updateColumns the column data (name => value) to be updated if they already exist.
+     * If `true` is passed, the column data will be updated to match the insert column data.
+     * If `false` is passed, no update will be performed if the column data already exists.
+     * @param array $params the binding parameters that will be generated by this method. They should be bound to the DB
+     * command later.
+     *
+     * @throws Exception|NotSupportedException if this is not supported by the underlying DBMS.
+     *
+     * @return string the resulting SQL.
+     */
+    abstract public function upsert(
+        string $table,
+        Query|array $insertColumns,
+        bool|array $updateColumns,
+        array &$params
+    ): string;
 
     /**
      * Contains array of default condition classes. Extend this method, if you want to change default condition classes
