+++ src/Schema/Schema.php
@@ -414,48 +414,6 @@ abstract class Schema
     }
 
     /**
-     * Unquotes a simple table name.
-     *
-     * A simple table name should contain the table name only without any schema prefix. If the table name is not
-     * quoted, this method will do nothing.
-     *
-     * @param string $name table name.
-     *
-     * @return string unquoted table name.
-     */
-    public function unquoteSimpleTableName(string $name): string
-    {
-        if (is_string($this->tableQuoteCharacter)) {
-            $startingCharacter = $this->tableQuoteCharacter;
-        } else {
-            $startingCharacter = $this->tableQuoteCharacter[0];
-        }
-
-        return strpos($name, $startingCharacter) === false ? $name : substr($name, 1, -1);
-    }
-
-    /**
-     * Unquotes a simple column name.
-     *
-     * A simple column name should contain the column name only without any prefix. If the column name is not quoted or
-     * is the asterisk character '*', this method will do nothing.
-     *
-     * @param string $name column name.
-     *
-     * @return string unquoted column name.
-     */
-    public function unquoteSimpleColumnName(string $name): string
-    {
-        if (is_string($this->columnQuoteCharacter)) {
-            $startingCharacter = $this->columnQuoteCharacter;
-        } else {
-            $startingCharacter = $this->columnQuoteCharacter[0];
-        }
-
-        return strpos($name, $startingCharacter) === false ? $name : substr($name, 1, -1);
-    }
-
-    /**
      * Returns the actual name of a given table name.
      *
      * This method will strip off curly brackets from the given table name and replace the percentage character '%' with
